Progress on Assignment Info system, comments
Also started to think about the design of a configuration system.
This needs to be done before continuing the Assignment Info system
since the config will be used to construct the other structs

I should add a debug stack trace which can print out a trace of the stack on
exit calls, segfaults (signal programming here), timeouts, etc

It may be worthwhile to design a special utilities library for multiple
languages but obviously C++ being the priority since my personal projects have
generally been in the aforementioned language.